Extract cabal cradle out of Cradle.hs
Cradle.hs grew a lot and is becoming to be unwieldy.
We fix this by extracting the most complicated cradle (cabal) into a
separate module.
To make this possible, we had to additionally extract a couple of types
and functions into separate modules as well to avoid cyclic module
dependencies.

This commit should not change any behaviour, but merely adds
documentation and moves definitions and data types.
